refactor: relocate classes to appropriate application directories
